moved function across files
